puppet: unset GIT_INDEX_FILE env var in post-commit hooks
We run a systemd unit/timer that uses gitpuppet to periodically
commit to the private repo (only one one node, not all
puppetmasters/puppetservers). We have been having the following
issue: most of the times, after a commit to /srv/private
on puppetmaster1001 (the current active host), we noticed
a staged revert under /var/lib/git/operations/private. The latter
is updated by the former's post-commit hook, doing git pull
as post-commit hook. The main issue seems to be the GIT_INDEX_FILE
env variable, that is populated for the /srv/private repo
and that gets carried over in the post-commit hook, somehow making
the git pull under /var/lib/git/operations/private inconsistent.

I tested this issue on a local simplified repro use case, and
unsetting the env variable fixes the issue.

Bug: T368023
Change-Id: I39d40e62823a9960b584656a2c663f1d68a85d18