dynamic proxy: Add an endpoint for scrubbing out nonexistent backends
We can no longer do this with precision in designate-sink since
designate doesn't know what IPs are removed when a VM is removed.
Instead, this function validates each proxy route in a project
and discards routes that point to invalid IPs.

This will also purge any proxies with a frontend and backend in different
openstack projects.

Bug: T371707
Change-Id: I6364ea19b2babf5f768ff7fdafca2cefe5b51348