hiera:benthos: remove Benthos from ulsfo using benthos module
With 1059265 is now possible to fully remove Benthos using the
profile::benthos::ensure: absent key. To achieve this a hiera
configuration with the instances to be removed needs to be present the
host(s). These will be removed in a subsequent commit.

Hosts: cp4037.ulsfo.wmnet,cp4038.ulsfo.wmnet,cp4045.ulsfo.wmnet
Bug: T371492
Change-Id: I079bb42f869f95af31f9e5f1791d89c26e190266