Add mtr to standard packages for WMF hosts
MTR is a very useful traceroute tool which can help quite a lot with
troubleshooting networking issues between hosts on the network.

It's a small package of about ~70KB, and would be useful to have
installed by default on our hosts imo.  This patch does just that,
and removes it from other specific profiles where it had been
included previously.

Change-Id: I02c74ec09c7e35099ce66c542b861efaee21c85e