ci: replace ferm::service with firewall::service for jenkinsagent
Another case where we just do the first step to move away from ferm
to a generic class that can be used for BOTH, ferm or nftables.

Noop but removes a blocker to switch in the future.

Also the code is simplified since "srange" needs to be an array now
and the DNS lookups happen automatically.

So we can simply pass the array of hosts from Hiera and be done.

See result in compiler:

https://puppet-compiler.wmflabs.org/output/1060483/3587/contint1002.wikimedia.org/index.html

Bug: T370677
Change-Id: Iecc9bc852b50cca818acfb22368564273b068e1d