Add a matomo_plugins component to the apt private repo
We need to distribute a non-free package using our private apt
repository to the server running Matomo. This package has been built
locally using the configuration files here:

https://gitlab.wikimedia.org/repos/data-engineering/matomo/plugin-customreports

Before adding this to the private repository, we need to add a new
matomo_plugins component to the repository and configure the apt sources on
the matomo host to access it.

Hosts: O:apt_repo,O:sretest,O:matomo
Bug: T370203
Change-Id: Idf59a946f3f44c3f64445eedc82b9ece10c07f16