apt-staging: Add check for packages in protected branches
Not every branch will match '-wikimedia', so add a check to see if it's
built on a protected branch as well. We could consider deprecating the
branch name check in the future if we decide that a protected branch
check alone is sufficient.

Also adds more debug logging to make it easier to find out why a package
isn't selected for download, and adds the option to specify a regex for
a build creation job

Change-Id: Id16b36526fc5e975a6e0f5500da62b488c8be6f7