varnish: Remove carriers netmap
We no longer use carriers.json, which was previously used for comparing
the X-Carrier header values with the file. This functionality was
removed with da1b4c74 and the remnants here are leftovers.

"X-Carrier" contained <mobile-country-code>-<mobile-network-code>
entries and "X-Carrier-Meta" contained the zeronet metadata like "wap".
Combined, these created the "X-CS2" header.

Bug: T370200
Change-Id: I0bdf85e736701e1728027a991fd2e2f9ddfc080f