Netbox: set RQ_DEFAULT_TIMEOUT back to default of 300
RQ_DEFAULT_TIMEOUT default is 300, so removing the current set value as
it was a wrong temptative to fix a timeout issue. Better to have it fail
faster as no job should take more than 5min.

If needed we could investigate using RQ_RETRY_MAX as well, but this
risks hiding the issue instead of solving it.

Bug: T341843
Change-Id: I4692daf12ce0ae5959fc76727758cdc58ab823e1