cephadm: separate templates for zonegroup setup and rgw placement
We need to specify the endpoints for each RGW zone (so that multisite
knows how to find the other zone correctly); unfortunately if you
specify zone_endpoints then cephadm does not start up any RGWs, so you
need a separate spec file to deploy them.

When setting up a secondary zone, the admin might need to briefly add
a realm_token to the zone_spec, so set show_diff to false for this
file so that the credential won't leak out in puppet change output.

Bug: T279621
Change-Id: Ide5835781ffa06202c2c343c6f9de6c5707c8fe4