confd: fix error state file name in check
In T363924, the error state file written by the confd-lint-wrap script
was updated to match the unambiguous name of the underlying template,
rather than mirroring the name of the rendered template tempfile.

This was done in order to address spurious alerts when two resources
have identical basenames (i.e., stored at different paths), an issue to
which this check was also susceptible.

The fact that this was not updated a couple of months ago was an
oversight.

Interestingly, the only place where this check is currently used is in
mw-cli-wrapper.py. However, that would never have worked anyway, since
the mediawiki.yaml confd::file resource in P:conftool::state configures
no check command.

Bug: T363924
Change-Id: I7a39bfb48d121d9d136e7a9d5bf969b93bdbfb71