Disable wiping LVM signatures on cephosd server reimages
We found that when reimaging cephosd servers, the existing LVM
signatures were wiped from the disks, which meant that they could not
detect the existing OSDs on the disks.

Puppet then create new OSDs instead of activating those which were
already present on the disks.

This change ensure that d-i no longer overwrites LVM partitions
discovered during the installation.

Bug: T372783
Change-Id: Ic2e361917d91f245620f0c51551bcb03f68b4144