site: apply gerrit role on gerrit2003
Using the temporary "insetup::gerrit" role and profile we have step
by step tested and installed gerrit on the new hardware.

Now we arrived at the step where it's identical to the normal production
role and we can prevent that the service IP is bound.

This means we can apply the normal production role.

Bug: T372804
Change-Id: I218b9f3d9c4dd9b92bb49831ed82663fbac34c0a