D:apereo_cas::service allow exclusion of LDAP groups.
When using the CAS protocol, all the users LDAP groups will be sent to
the client. For users with many groups this might results in "tokens" to
large for Apaches token storage.

The most relevant case is to filter out tools related LDAP groups, as
these would not be used to allow access, but potentially take up a large
amount of space in the tokens.

Bug: T369205
Change-Id: I0b6280b73a85a207d3055a23af780754ff44a733