gitlab: add replica hosts to Icinga
In response to the comment at I946dd0b73b6be2d how the cookbook
fails due to gitlab replica hosts not being found in Icinga.

Add array of replica hosts to Hiera and add a parameter
to pass them to @monitoring_host which will add them
as a "custom host" (service name) to Icinga.

This is the same way it's been done for other services
that aren't the primary host name, like phab.wmfusercontent.org.

Bug: T363564
Change-Id: I6112339578244cc9b48bed207bb0babc59a1bd20