cephosd: remove a subset of LVM signatures during reimage
We have seen problems during reimages of cephosd servers, whice are
caused by either *all* of the LVM signatures being removed, or *none* of
the LVM signatures being removed.

We would like to be able to remove only those LVM signatures that are to
be used for the O/S installation, whilst leaving those that are assigned
to Ceph OSDs untouched.

In order to do this, we:
* Find out if there is a PV signature on a software RAID device
* If there is such a PV:
 * Remove all LVs held on that PV
 * Remove the VG matching the hostname-vg pattern
 * Remove the PV itself

Then we tell the installer not to remove any other LVM signatures.

Bug: T372783
Change-Id: I826fd5b5c3c2250147b11aecb4fd1f4f40a99acd