gerrit: fix todo from 2022, remove nist key setting
Per the TODO comment this should be removed once the
relevant ticket is resolved.

This already happened in 2022.

This was added and explained in I5c9318727f10a562c.

We are now at Gerrit 3.10, way newer than 3.6.

What this actually does is:

+KexAlgorithms curve25519-sha256@libssh.org,diffie-hellman-group-exchange-sha256

in the sshd_config.

The MINA version as of today is APACHE-SSHD-2.12.0 per "telnet localhost 29418".

If replication over ssh between Gerrit servers is still working after
this can be checked in /var/log/gerrit/replication_log.

We might also have to edit:
/modules/gerrit/files/homedir/.ssh/config in the repo though.

Bug: T315942
Change-Id: Ie1ee6bb281dd1f9067f8f00f3f3c346ca92777ab