releases: upgrade Java JDK version from 11 to 17
As reported by Bryan Davis, jenkins on releases hosts
is whining about the Java version becoming EOL soon.

The releases hosts are already on bullseye but the
Java version is specifically set to 11 in Hiera.

As far as I remember this was purposefully made separate
when we upgraded the machines to bullseye.

Seems like it's time to do this step now.

Bug: T359795
Change-Id: Iefc86b35a03beb71322c7ec3d5073055abbd1171