cephosd: Assemble the MD RAID arrays, so that they can be removed
The most recent reimage attempt of a cephosd server resulted in a
failure because /proc/mdstat had not been populated at the time when the
script was running. This meant that it was not possible to delete the
RAID metadata, nor were the LVM signatures on those MD devices able to
be deleted.

This change causes the script to run `mdadm --assemble --scan` early in
the script, which assembles the arrays.

We can then loop over /proc/mdstat and stop each array and then wipe the
signature. Removing the LVM signatures is then unnecessary.

Bug: T372783
Change-Id: I9b4d06394fd79ca4219501ff53a3910eafaee991