cephosd: Remove LVM signatures in addition to MD RAID metadata
The most recent reimage attempt successfully deleted the MD RAID
metadata, but when it was recreated the original LVM signatures were
still present and were used.

This commit causes the script to delete the LVM signatures for the O/S
volumes, if they are present when the MD RAID devices are assembled.

Bug: T372783
Change-Id: I8d4a6d23994acf8e7a759184d53b6f1956af8524