apt-staging: Remove 'staging' flag from distribution name
We build packages for 'bullseye-wikimedia', but when we set the
distribution in reprepro to be '-wikimedia-staging', it won't add
packages.

There seems to be an override option, but that's on a per-package basis,
which seems clunky. Since the staging and non-staging repos are kept on
separate hosts for the moment, it seems simpler to set the
codename/distribution to be the same on both hosts, and simply
distinguish staging vs non-staging by changing host where the package
comes from.

Change-Id: I8d49605bf10fb4dd86472fed85e099e1df55b06a