profile::firewall::nftables_throttling: fix issue of global metering
The definition in the throttling chain is metering/rate limiting in
the wrong place. This leads to global metering and throttling and
blocking of all IPs as soon as one client exceeds the thresholds.

This patch should fix the metering issue by introducing another set used
for the actual metering. As soon as a client connects the IP is added to the
metetring set. If the client exceeds the threshold, the IP is also added
to the DENYLIST.

Bug: T366882
Bug: T365259
Change-Id: Ib4a299d586337018571e1d42028e3049a0226c5c