prometheus/gerrit: also add size of tracking list to exporter
Next to the size of the actual denylists, also add the size
of the tracking lists to the prometheus exporter data.

This does give us data about IPs being blocked; it just gives
us the general amount of users on Gerrit.

But nevertheless it seemed interesting enough to add it to
the dashboard as well.

Bug: T373136
Change-Id: If7de91dfd6d7abf01c371c0f7af06a1ee72a2d54