archiva: allow trailing slash for top directories
In T373352 I encountered a Java build failure for a missing dependency.
I used my browser to head to the URL used for the Archiva repository:
https://archiva.wikimedia.org/repository/mirrored

That yields a 404. I have thus assumed Archiva has been entirely
decommissioned which led to some confusion. The reason for the 404 is
23fdde6e8506fb7a0ba872efba61afae40bc5300 which restricts requests to eg
`/repository/mirrored/.*`, it was made to require a trailing slash when
the previous regex did not.

Prefix the list of allowed characters with a / and makes the whole
optional. This allows requests made to the top directory to not have a
trailing slash.

Bug: T359031
Bug: T373352
Change-Id: I8217e6a134a6fd8aed110875a33acda24ec54925