deployment: Add undocumented flag --helmfile to mwscript-k8s.
For testing, it's helpful to work from a locally modified
deployment-charts checkout. In order to that I've also been copying the
script to change the hardcoded helmfile path -- at this point it might
as well be a command-line flag. Maintenance-script runners won't need
this, so it doesn't need to clutter the --help output.

Change-Id: I4694bfa4bfc8fceb8a10046f207d539b6db848b6