releases: drop java_home parameter for releases/mediawiki
follow-up to Iefc86b35a03be

Per the plan, drop the java_home variable, since it's not actually
set in the puppet repo but in the jenkins-deploy repo in gitlab.

Not passing it anymore to the jenkins class means it will fallback
to /usr/lib/jvm/java-8-openjdk-amd64/jre but that should be irrelevant.

We just want the openjdk-17 packages to be installed and then the home
will be changed in the other repo.

Bug: T359795
Change-Id: I15fedff613cf7e1cb58f5ae696af4a9b463ab109