LVS: Only allow IPv6 default route from RAs on primary interface
Adjusts sysctl's so that 'net.ipv6.conf.default.accept_ra' is set
to 1 for all ints, ensuring that IPv6 address autoconfiguration
will take place, but settting 'net.ipv6.conf.all.accept_ra_defrtr'
to 0, so by default RA processing on an interface doesn't cause
a default route to be added.

Then we explicitly enable accept_ra_defrtr for the primary interface
so that we get the required default route only on that one.

We do this only on the backup LVS in codfw, lvs2014 to start with before
rolling it out everywhere else.

Bug: T358260
Hosts: lvs2014.codfw.wmnet, lvs1020.eqiad.wmnet, lvs4010.ulsfo.wmnet
Change-Id: I574b55177ef6d11c1524110a6bcc145c8fbc155d