profile::puppetserver: set java_start_mem to 40g
From the Puppetserver dashboard it seems clear that pre-allocating
a 48g heap (min==max equals to pre-allocation of the entire memory
size set) is overcommitting, since the used memory on average is
much lower. We have some scenarios in which we end up in OOMs due
to the OOM killer, so relaxing the start memory size could help with
that.

Bug: T373527
Change-Id: Ifc1b93fe9987602fe8b8118102ec90858aa7f542