loki: increase chunk flush interval
Why:
A log stream is a log event that shares the same set of labels hashed
together.  A chunk is a time-delimited repository for a log stream.
When a log event for a stream is ingested, a new chunk is created in
memory.

chunk_idle_period is 30m by default.  This is how long chunks
should sit in-memory with no updates before being flushed, and a new
chunk created in memory, if they don't hit the max block size.

max_chunk_age is 2h by default.  The maximum duration of a timeseries
chunk remains in memory. If a timeseries runs for longer than this,
the current chunk will be flushed to the store and a new chunk
created.

What:
Set chunk_idle_period and max_chunk_age to 168h (one week) to reduce
the overall number of chunks flushed to disk for sparse log streams
(alerts).

Hosts: grafana1002.eqiad.wmnet
Bug: T335610
Change-Id: I2d475854d57df4d8ec86fab18c105771c4ee1681