trafficserver: Allow XWD to be used for wikitech
Why:
In the spirit of making wikitech yet another wiki (YAW),
we would like to test our progress on mwdebug* hosts.

What:
Currently, traffic to wikitech is directed to labweb
regradless if XWD is set.

Warning:
By merging this patch, when someone tries to access
wikitech.wikimedia.org with XWD on, they will get
an error, as this vhost is not configured on our
current apache configuration.

Hosts: cp1100.eqiad.wmnet
Bug: T371537
Change-Id: I8093f411d51d1cd8e8fbf8f2a44b2ed41d7c5519