Temporarily disable stunnel for the Puppet 7 migration of deployment hosts
We want to initially only move the secondary deploy host to Puppet 7
and stunnel relies on the Puppet CA instead of cfssl, as such
stunnel-encrypted tunnels between P5 and P7 hosts aren't going to work.

Follow the approach that we used for the grafana hosts/loki and disable
stunnel until both deployment hosts are on Puppet 7.

Hosts: deploy1003.eqiad.wmnet,deploy2002.codfw.wmnet
Change-Id: I048512c42fba6574876f73c3e458405f991d0e03