Lower the number of slots that the enwiki dump uses
We have recently seen issues affecting the core database server
lag, which are tied to the dumps architecture. Most recently, the enwiki
dump that runs on snapshot1012 has been implicated in these incidents.

We have decided to reduce the maximum number of parallel dump process
that run on this host from 28 to 18 as a means of mitigating this
overloading of the s1 section replica.

Hosts: snapshot1012.eqiad.wmnet,O:dumps::generation::worker::dumper
Bug: T373904
Change-Id: I6b7ca03b394351ab3adce5fb35e7e07652dcdae5