[Analytics][Refine][test-cluster] Bump test cluster Refine job version
As part of migrating the Refine job to Airflow
(from systemd timer), we need to ensure the new
process produces the same output as the old one.

The legacy Refine version differs because:
* The deduplication function was not
  deterministic (not consistently removing the
  same duplicates).
* The geocoding function needs updates to match
  the legacy output.

This change updates the Refine version for the
analytics test cluster. A separate patch will
handle the production environment later.

Bug: T356762
Change-Id: I4423af80848bf35458c768c9e1ab2308b55679ca