profile::firewall::nftables_throttling: throttle connections not packets
This change refactors the nftables_throttling profile to throttle based
on parallel connections instead of the number of new TCP connections.

As outlined in T365259#10116300 Gerrit creates quite a lot of new TCP
connections when pulling a repository with submodules. This would trigger
the existing throttling implementation.
So let's try to throttle based on parallel TCP connections. The default
is 32 parallel connections for every IP (for both Gerrit and GitLab).

The TRACKINGLIST has no timeout anymore because it only tracks active
TCP connections. IPs are removed as soon as there is no open connection.

Bug: T365259
Change-Id: I41bafea13f304866ccf666a67fbfce0a38f07fd5