exim: fix VERP handling for mediawiki
The bounce handler was not working correctly since we've switched it to
use TLS to connect to mediawiki. This is due to the very brittle way
MediaWiki handles external IP addresses, which requires us to inject the
X-Client-IP header into the request so that REMOTE_ADDR is set to the
IP of the MX server and the bounce is not rejected.

Bug: T338761
Change-Id: I7516f58141df027c1378ad3763ba3d4951cc207e