Add the anycast VIP for radosgw to DPE Ceph servers
We are going to be using anycast to support load-balancing of the Rados
Gateway services on the DPE Ceph cluster.

This requires adding the bird::anycast profile, along with a suitable
configuration for advertising addresses.

The addresses have already been assigned in netbox and are active.

Hosts: O:ceph::server
Bug: T330153
Change-Id: I39cbce1be1270636080730aa3869df15e960b2ca