mediawiki::maintenance: Remove php72 remnant
Why:
This component hasn't been used since we moved to php 7.4. The reason
for this is that both the php 7.2 and php 7.4 components have a higher
priority than the default buster one. Between the 2, php 7.4 has a
higher version number and is preferred over the 2 anyway, even without
other php things depending on it

What:
Remove mention of the old php72 component

Hosts: auto
Change-Id: I9b852d7357528f5d560768df7531a565d9e427de