puppet8: ensure cassandra passwords are defined
Prior to this change the passwords::cassandra class was included in the
cassandra class. This meant that the variables defined in
passwords::cassandra were not available when looking up the class
parameters for cassandra. It is unclear why this did not break on Puppet
7. I also tried using chaining arrows and the require function, but
neither appear to work for variable assignment.

Hosts: P:cassandra
Bug: T372664
Change-Id: If3e8f6d6d1dc9da65afd0c63333ac9dfcbb8eeec