planet: drop firewall rule for http from localhost
First there was just an apache on port 80 and traffic
to it was allowed from anywhere, no srange.

Then an srange was introduced to lock it down to just from CACHES Icba8af083f3f5a82b.

Then envoy was introduced to terminate TLS and port 80 srange changed to
localhost instead of CACHES.

As pointed out on I84c8bec01991 this is probably pointless now since
traffic from localhost is allowed anyways.

Change-Id: Idab22dd3700378e49791ae328f8b9e5ad92a9855