lists: Set number of processes for mailman3_runner to minimum of 14
Currently it's set to alert unless there are exaxctly 14 processes.
After today's reboot, it started alerting because there are 15 processes
now, including two for `--runner=rest`. The original alert seems to be
checking for crashed runners that never get restarted, but having a
minimum number of runners would seem to mesh with that.

Change-Id: I9f188585a317f0be8a7043bc1c4ddf58ece537db