ganeti: Install bridge-utils on Bookworm
Up to Bullseye one of the Ganeti packages (specifically ganeti-3.0) had
a dependency on bridge-utils. This got dropped between bullseye and
bookworm since ganeti internally uses iproute2. 

However, we currently use ifupdown to configure network interfaces and if ifupdown
uses bridge-utils to configure bridges (and fails with a cryptic error if the latter
isn't installed). As such, explicitly install it via Puppet for Bookworm.

Don't install it for the routed Ganeti setup case, we don't need it there.

Change-Id: I47f8b89f7e1ca121c4850ef538a65ecfbf867435