lists: Mask mailman3 service on non-active host
The active host will have $service_ensure set to present, so ensure that
it's unmasked on that. On other hosts, set the systemd unit to masked so
that the mailman3 process never attempts to start. This is because the
process will never start up cleanly due to no grants on the mysql server

Change-Id: I7dd1cd4c67adccfdd8bfc73d5be6ff433e964222