P:trafficserver: extend x-wikimedia-debug-routing for mwdebug-next
The mwdebug-next service is now up (T372604) and will soon be ready for
testing PHP 8.1. In advance of this, add support for routing requests to
mwdebug-next services on k8s via x-wikimedia-debug.

Given that this makes the debug_map table more costly to construct, this
change also moves its declaration into the top-level scope of the script
so that it's constructed once on load.

While this script is of course not used on a performance-critical path,
this still seems like a sensible improvement.

Bug: T372605
Change-Id: I95f3b900885e30e97f0d4b35067f07fe018bc49b