Manually define BGP neighbors for cephosd1*** Anycast BGP
The current puppet bird class is defined in a way that it will
use a host's default route (i.e. top-of-rack switch) as BGP
neighbor, unless profile::bird::neighbors_list is defined
in hiera.

This works decently well for the existing use-cases.  We define
this array at the site level for eqiad, codfw, ulsfo and eqsin,
where we have L2 switches and the hosts doing Anycast BGP
peer with the CR routers.  At sites where we only have L3
switches, like magru/drmrs/esams, we do not define this var
anywhere and the role causes the top-of-rack to be used instead.

While working on this for the ceph hosts it turns out there is a
problem, as they are in eqiad rows E/F, on per-rack vlans which
prevents them peering with the CRs in BGP, but as hiera has
this set at the site-level they are being configured to do just
that.

This patch overrides the site-wide var to set the correct switch IPs
for BGP for these hosts.  Ultimately this should be a temporary
solution, we don't want to have to define this everywhere, but I
will wait for Arzhel's return to discuss what way is the best here
from a puppet point-of-view.

Bug: T330153
Change-Id: I8930572550b6708af3b8cc8b69e980e75580fa07