ntp: standardize the use of ntpsec in the configuration as well
We are already running ntpsec in bookworm and while the service is
aliased and the initial configuration copied, make sure we standardize
this in the Puppet configuration as well. Only DNS hosts are using this
so should not be an issue.

There is some more cleanup so that there are no errors in the ntpsec log
but that's for another commit.

Hosts: dns6001.wikimedia.org, dns1004.wikimedia.org
Change-Id: I0beda0df51c512d78c5302e3fb88c9fb9456b4aa