Reduce airflow-analytics log retention from 90 to 60 days
Recently we have seen a marked increase in the size of the airflow logs
on an-launcher1002, particularly since introducing dynamic DAGs for
canary events.

Logs are purged from the filesystem by a systemd timer that runs daily.
which is set to retain 90 days' worth of uncomporessed logs, by default.

We are unable to add more disk space to this server, so in order to
reduce the contention for disk space, we wish to reduce this retention
time from 90 to 60 days.

The project to migrate Airflow to Kubernetes is well under way and we
will be using S3/Ceph based logging for the new instances. We can review
the retention period and other settings at that time.

Hosts: an-launcher1002.eqiad.wmnet
Bug: T370437
Change-Id: I6097685a2c076090c244547546b75c3979458f99