Change config for cephosd's in eqiad to peer with switch global addr
I had used the link-local here, to match what we do when we auto-
discover it based on the default route.  But for some reason
BFD won't work with that scenario.  Difference is in Eqaid the
peering is in a VRF, but other than that not sure why it doesn't
work.  Anyway it does work ok with the global address used both
sides, which is ok for now as we are specifying the neighbors
manually.

Bug: T374379
Change-Id: I367def95ae369cdf5f5001d069416c1f9d715e49