Use global unicast to peer from cephosd but allow LL for BFD in
It appears that there is a problem on our EVPN-based switches -
where we configure networks hosts are connected to in a VRF /
routing-instance - properly establishing a BFD session to
hosts.  On the face of it the hosts send the BFD init messages
to the switch link-local IP, but the switch does not process it.

Establishing the BGP from the host to the switch's global unicast
IP does work, and we can get BFD to work in this scenario.  With
the switch BGP peer configured as the global address the host
sends the BFD messages to it also, and the switch processes them.

To complicate things further (of course), the switch still chooses
its own link-local IP to send BFD messages to the host.  Which get
dropped by the host's firewall as they only allow from the
configured peer address currently.

So this patch modifies the setup to use the global IPs for the
networks the cephosd devices are connected to, but also adds a
rule to allow BFD from the link-local to accept the switch
messages.  This should be considered temporary, ultimately
using global unicast both sides and for all protocols is
probably better but this should unblock for now.

Bug: T374379
Change-Id: I09ed10dda5f74bcb7e413000ad4a33d14a7dc472