errorpage: Include request ID early in HTML source
Some users of the API will only log the beginning of the response body
if it cannot be decoded in the expected format (usually JSON) – an
example for this can be seen in T374344#10132077, from a Python tool
using the mwapi library. Let’s include the request ID in a comment, as
early as reasonably possible (but after the charset and <title>, both of
which are arguably more important), so that such cases can be debugged
more easily.

Follows up Idb96796483, which added it to the HTML body.

Bug: T291192
Change-Id: I8cdbdbee6ece1550ca81377b2c3fab92ab52d887