Trust DSCP markings from VMs on routed ganeti hypervisors
We use the postrouting chain in iptables/nftables to mark the DSCP
bits in packets if we want them to get special treatment by the
network QoS configuration.

On legacy ganeti hosts we don't need to do anything special, as
packets traversing the linux bridges do not go traverse the iptables
chains defined.  However in routed mode they do traverse the
postrouting chain, and we need to avoid re-marking or un-marking
packets that had DSCP set inside the VM.

Adds a rule to ignore such packets based on input interface name
of tap*.  Tested locally with dcl and works as expected.

Bug: T374392
Change-Id: I2ac0edcb360a09765a667436a682bc4c113c2ab9