role::puppetserver: add TLS+HTTP stack to publish SHA1 values
The puppetmaster stack uses profile::configmaster to expose
the SHA1 values produced by puppet-merge.py to the config-master
nodes (that currently mod-proxy to puppetmaster1001.eqiad.wmnet
to fetch those values). The SHA1s are used in various way, one of
them most important ones being git-sync-upstream (used everywhere
in cloud to keep prod-puppet and labs-puppet in sync).

In order to allow puppetserver hosts to let puppet-merge to run,
we need to re-create a similar setup in which we are able to
expose the SHA1 values when they are produced. Instead of using
profile::configmaster, that would bring in a ton of resources not
needed, create a simple TLS + HTTP stack that does it.

Bug: T374443
Change-Id: Ief5ed5ea150911f61a7e0c9bc85c37508f3fd968