config_master: Explicitly configure the server for Puppet merges
config-master nodes proxy to the Puppet server on which Puppet changes
are merged (currently puppetmaster1001). However, the Puppet code reuses
the global setting for the puppet CA, which is confusing since the Pupppet
merges are totally unrelated to where the Puppet CA runs.

Instead, disambiguate things and designate a specific Hiera setting which
specifically refers to the host from which puppet-merge can be run and 
adapt the config-master template to use it.

With this in place, when we switch puppet merges to puppetserver1001 we'll
need to change the Hiera entry so that config-master proxies to the new
host.

Hosts: C:profile::configmaster
Bug: T366355
Change-Id: If2d6798dbb958b36c32d5cc4673ba89404b15ddd