P:ntp: drop backward compatibility for ntp (only use ntpsec)
The ntpsec configuration while backward compatible (to some extent) with
the older ntpd configuration still results in some error messages in the
journal that we should probably clear up. Since we have switched to
ntpsec everywhere, this should close that loop.

List of changes:

1. CONFIG: peer deprecated, treated as server: dns1005.wikimedia.org
We have removed peer and are now using server.

2. CONFIG: restrict nopeer ignored
nopeer has been deprecated, so removed.

3. CONFIG: restrict notrap ignored
There is no trap support in ntpsec, so notrap doesn't apply, removed.

4. statistics directory /var/log/ntpsec/ does not exist or is unwriteable,
We can create the directory but leaving this for now as we don't use the
statistics anyway.

5. We are dropping iburst as we feel that the (only) benefit of faster
sync outweighs the concerns we have around it affecting accuracy. We are
removing this from our own servers and the public pools.

Hosts: dns6001.wikimedia.org, dns1004.wikimedia.org
Change-Id: Id62f3bf2a4d11d6c8d442f604dfabb964a22c20c