vrts: switch inactive host vrts2001 to nftables as firewall provider
This is like If30ea3b814d813 but first only doing it for the inactive
host before moving the same settings to the role level that affects
all hosts.

Since we are currently in the middle of setting up new hardware
we could also use vrts1002/2002 but they don't have the puppet
role just yet.

Bug: T370677
Change-Id: I92a78bbf2f02b6a1e7ab221216c3127c1fa308e6