envoyproxy: add route-level idle timeout for tlsproxy
Add an idle timeout for inactive connections at the route level. We
have noted that this was required on the service proxy and envoy
sidecar level previously, and this also needs to be set at the service
proxy level. This mirrors the existing `upstream_idle_timeout`
parameter in the mesh.configuration module in deployment-charts.

Also add support for the existing envoyproxy downstream_idle_timeout
configuration option to tlsproxy.

Bug: T373517
Change-Id: I7453618cfeb7fc06fd9461c67e6c8264b9a5964d