Move the misc_crons dumper role from snapshot1017 to snapshot1016
We need to carry out a reboot of snapshot1017 but we would like to do so
without interrupting any of the dumps that are carried out on this host.

We will therefore move the dumps::generation::worker::dumper_misc_crons_only
role from snapshot1017 to snapshot1016, which is a testbed host.

Hosts: snapshot1016.eqiad.wmnet,snapshot1017.eqiad.wmnet
Bug: T366555
Change-Id: I85d46e429f6f92b389ce99dbe07cd21bcf3ea796