profile::docker_registry_ha::registry: add sslcert::dhparam
For some reason when I bootstrapped a new VM on Bookworm (registry2005)
nginx started to look for /etc/ssl/dhparam.pem, but the file is
not present on disk. The file is present on the other registry nodes,
but I am wondering if over time some TLS-related refactoring left
this use case half-broken. Explicitly add the class that renders
the file to the registry's profile.

Bug: T374928
Change-Id: I9e3ad7d79532f1b707cb98c43caa518bf1f3946a