Add an hdfs_file type and provider
This commit adds support for managing files on HDFS by the use of a new
resource type called `hdfs_file`. It uses the hdfs-rsync utility to sync
the files from a temporary local copy to a given path on HDFS.

It uses the `hdfs` cli to remove the file when ensure => absent is
passed.

As per the posix file provider type, it supports either the content or
source parameters, but not both.

The user, source, and mode parameters are also required. The hdfs user
account is used, which has the permissions required to change the owner
of the file.

Bug: T323692
Change-Id: Ida5984ae0098aad91eaf2fa0717f08e12e215c43