Permit db1208 to access the Ceph/S3 endpoint
We wish to use this host to backup several of the S3 buckets to its
local disk, so that they may be added to the bacula configuraiton.

In order to do so, we have to permit access to the S3 endpoint via the
firewall on each of the cephosd servers.

Hosts: O:ceph::server
Bug: T372908
Change-Id: I8e8858dd3c260a6f487d9bc2716a92bcec487f5d