profile::puppetserver: correctly populate the MASTERS env variable
In the puppet-merge.sh script we have two env variables:
- MASTERS
- WORKERS
The first one was meant to target only puppetmaster frontends,
meanwhile the latter all servers (including backends). We then added
puppetservers, and the difference between them got blurry.

We moved the puppet-merge script to puppetserver1001, and now
MASTERS is only relevant to understand on what nodes the labsprivate
changes are propagated. It is currently too broad, and we got reports
of puppet-merge failures of labs-private changes, since they got
pushed to puppetmaster backend hosts that don't have the repo
checked out.

Reduce the scope of MASTERS for the moment, we'll get rid of it
when the puppetmaster decommission happens.

Bug: T374443
Change-Id: I85260f593e5b3870aeb6913d5012120ae6f2d69d