hieradata::services_proxy::envoy.yaml: fix duplicated port
It seems the data-gateway and the shellbox-video services
are configured with the same port.
We suspect this is the reason the data-gateway is not accessible
from the commons-impact-analytics service pod.
This change modifies the data-gateway port,
since we believe it's not used anywhere else than in
commons-impact-analytics.
It sorts the order of the shellbox-video service,
since it was not port order.

Bug: T368035
Change-Id: I1472e12afed685b75c36b7f7dd6b7072824225eb