types: extend Profile::Mediawiki_deployment
Updates Mediawiki_deployment to support the new release config mapping
described in [0]. In effect, we're transitioning:

  - namespace: mw-web
    release: main
    canary: canary
    mw_flavour: "publish"
    web_flavour: "webserver"
    debug: false

to the equivalent:

  - namespace: mw-web
    releases:
      main: {}
      canary:
        stage: "canaries"
    mw_flavour: "publish"
    web_flavour: "webserver"
    debug: false

where additional releases can be added and optionally override the image
flavours.

The older 'release' and 'canary' fields are still supported by scap on a
temporary basis. Further, the two formats are exclusive, in that a given
namespace must use one or the other (e.g., 'releases' and 'release' are
not permitted to coexist), and the only supported stage for the moment
is canaries.

A Variant type is thus a natural way to represent structural assertions
for the hieradata key at this time.

[0] https://phabricator.wikimedia.org/T370934#10104089

Bug: T370934
Change-Id: I7af6b2fffc6b48193ae1f72ef81c76f4a9882f33