git: add replicated_local_repo define
We sometimes have to handle data in production that fit a git repository,
but that either contains restricted/sensitive data or doesn't really have
a place in a public git repository.

Such data will need to be replicated across all servers that share the
same role. For all such situations, we have in the past come up with bespoke
solutions.

The conftool audit trail is one such example.

This is an attempt to standardize the process by creating a puppet define
that will create a git repository on the server, and then add ssh
keys/configuration to allow the specific user to push to the repository.

Finally, if a server pool is defined, the repository will be replicated
across all such servers using a post-commit hook. The limitation of this
approach is that rewriting history would break the process.

Bug: T374723
Change-Id: I9aea32c8d3f235bb2dbf920d70520b68b815cfc8